search active transition
