+++ components/menu/AppHeader.vue
@@ -1,17 +1,5 @@
 <script setup lang="ts">
-import { ref } from 'vue'
-import RoundButton from '../shared/RoundButton.vue'
-import MultiSelect from '../shared/MultiSelect.vue'
-
-const options = [
-  { value: 1, label: 'map-marker-outline' },
-  { value: 2, label: 'star-outline' },
-  { value: 3, label: 'wheelchair-accessibility' }
-] as const
-
-type Value = typeof options[number]['value']
-
-const selectedFilter = ref<Value>(options[0].value)
+import AppHeaderActions from './AppHeaderActions.vue'
 </script>
 
 <template>
@@ -25,17 +13,8 @@ const selectedFilter = ref<Value>(options[0].value)
     </h1>
 
     <!-- actions -->
-    <div class="mt-2 flex justify-between">
-      <RoundButton>
-        <mdicon name="magnify" size="24" />
-      </RoundButton>
-
-      <MultiSelect
-        v-model="selectedFilter"
-        :options="options"
-        type="icon"
-        class="w-[200px]"
-      />
+    <div class="mt-2 relative">
+      <AppHeaderActions />
     </div>
 
   </header>
